core: update to new modding api
