--- Blasphemous-Rando-Map.sln
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.32407.337
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RandoMap", "RandoMap\RandoMap.csproj", "{7751B777-DEB7-42D7-AD11-2893B3165634}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{7751B777-DEB7-42D7-AD11-2893B3165634}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7751B777-DEB7-42D7-AD11-2893B3165634}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7751B777-DEB7-42D7-AD11-2893B3165634}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7751B777-DEB7-42D7-AD11-2893B3165634}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B45CB2E4-3CE0-40C2-8558-D0423250B2D8}
-	EndGlobalSection
-EndGlobal
