+++ Blasphemous.Randomizer.MapTracker.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.8.34408.163
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Blasphemous.Randomizer.MapTracker", "Blasphemous.Randomizer.MapTracker\Blasphemous.Randomizer.MapTracker.csproj", "{AB3C490A-A787-49CB-8748-37104E613082}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{AB3C490A-A787-49CB-8748-37104E613082}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AB3C490A-A787-49CB-8748-37104E613082}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AB3C490A-A787-49CB-8748-37104E613082}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AB3C490A-A787-49CB-8748-37104E613082}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {E00FB6F4-CEAE-4A02-B766-9355BC86DC9B}
+	EndGlobalSection
+EndGlobal
