--- RandoMap/CollectionStatus.cs
@@ -1,15 +1,0 @@
-﻿
-namespace RandoMap
-{
-    public enum CollectionStatus
-    {
-        Untracked,
-        Finished,
-        AllUnreachable,
-        SomeReachable,
-        AllReachable,
-        HintedAllUnreachable,
-        HintedSomeReachable,
-        HintedReachable,
-    }
-}
