--- resources/keybindings/Rando Map.txt
@@ -1,1 +1,0 @@
-ToggleLocations: F1
